services/discord-bot/requirements.txt
# Discord bot dependencies
discord.py==2.3.2
aiohttp==3.9.1
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0

# Utilities
uuid==1.30
asyncio-throttle==1.0.2

# Logging
structlog==23.2.0

# Date/time handling
python-dateutil==2.8.2